Zmiana nazwy folderu projektu z 2-3 na 2-2
