+++ Zadanie-2-2/.gitignore
